# Potentially faster PropertyBinder