--- a/src/Serilog/Events/EventProperty.cs
+++ b/src/Serilog/Events/EventProperty.cs
@@ -97,7 +97,6 @@ namespace Serilog.Events
 
         public interface IBoundedPropertyVisitor
         {
-            void On(MessageTemplate parsedTemplate);
             void On(EventProperty property);
         }
     }