… test/Serilog.Tests/Support/Extensions.cs