This PR augments `PropertyBinder` to do not allocate the `EventProperty[]` and to directly pass the values to the caller, that requested it. I found two paths that are calling it:

- Logger, when creating `LogEvent` and then calling `.Dispatch`
- Logger, when calling `.BindMessageTemplate`

This PR is just a proposal. It requires some insight and potentially running more benchmarks.

### Benchmarks

I run benchmarks from `AllocationBenchmark` but it might require running the rest of the suite. 

```ini
BenchmarkDotNet=v0.11.5, OS=Windows 10.0.17134.829 (1803/April2018Update/Redstone4)
Intel Core i7-6700HQ CPU 2.60GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
Frequency=2531247 Hz, Resolution=395.0622 ns, Timer=TSC
.NET Core SDK=2.2.300
  [Host]     : .NET Core 2.2.5 (CoreCLR 4.6.27617.05, CoreFX 4.6.27618.01), 64bit RyuJIT
  DefaultJob : .NET Core 2.2.5 (CoreCLR 4.6.27617.05, CoreFX 4.6.27618.01), 64bit RyuJIT
```

#### Before

|               Method |         Mean |       Error |        StdDev |       Median | Gen 0 | Allocated |
|--------------------- |-------------:|------------:|--------------:|-------------:|------:|----------:|
|             LogEmpty |     11.85 ns |   0.3149 ns |     0.9085 ns |     11.82 ns |     - |         - |
| LogEmptyWithEnricher |     78.25 ns |   2.4580 ns |     3.8268 ns |     77.20 ns |0.0178 |      56 B |
|            LogScalar |    966.26 ns |  30.2759 ns |    82.8799 ns |    938.42 ns |0.1183 |     376 B |
|        LogDictionary |  5,912.98 ns | 406.5219 ns | 1,185.8435 ns |  5,570.99 ns |0.7095 |    2240 B |
|          LogSequence |  2,732.37 ns | 169.2209 ns |   482.7967 ns |  2,553.46 ns |0.2594 |     824 B |
|         LogAnonymous | 11,927.61 ns | 497.3753 ns | 1,450.8670 ns | 12,070.04 ns |1.0834 |    3440 B |

#### After

|               Method |        Mean |       Error |     StdDev |      Median | Gen 0 |Allocated |
|--------------------- |------------:|------------:|-----------:|------------:|------:|---------:|
|             LogEmpty |    12.23 ns |   0.3690 ns |   1.053 ns |    12.01 ns |     - |        - |
| LogEmptyWithEnricher |    87.15 ns |   3.4287 ns |  10.056 ns |    83.39 ns |0.0178 |     56 B |
|            LogScalar | 1,000.03 ns |  19.7406 ns |  27.674 ns |   993.39 ns |0.1049 |    336 B |
|        LogDictionary | 5,265.85 ns | 197.2796 ns | 562.850 ns | 5,133.75 ns |0.6943 |   2200 B |
|          LogSequence | 2,234.62 ns |  45.9707 ns |  68.807 ns | 2,210.84 ns |0.2480 |    784 B |
|         LogAnonymous | 9,715.81 ns | 118.8590 ns | 105.365 ns | 9,716.56 ns |1.0681 |   3400 B |