@Scooletz 2019-06-26
After seeing benchmarks, I can tell that this is not faster at all.