@Scooletz 2019-06-26
Next time will raise a PR when it's really better 🤦‍♂