@nblumhardt 2019-06-26
The approach was interesting - PoC's at any stage are welcome :-)